git-svn-id: svn+ssh://svn.code.sf.net/p/opennode2/src/trunk@1027 a46bb6aa-65bc-42c6-9d02-ca0d7e7f78c4

+++ dotnet/common-windsor-net/windsor-commons-core/FileUtils.cs
@@ -367,7 +367,10 @@ namespace Windsor.Commons.Core
         {
             try
             {
-                Directory.Delete(directoryPath, true);
+                if (Directory.Exists(directoryPath))
+                {
+                    Directory.Delete(directoryPath, true);
+                }
             }
             catch (Exception)
             {
